cov-script/pyproject.toml
[project]
name = "cov-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyghidra>=2.2.0",
]

[dependency-groups]
dev = [
    "ghidra-stubs==11.4",
]
